# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Sistem Pelaporan Prestasi Mahasiswa API
  description: |
    API untuk sistem pelaporan prestasi mahasiswa dengan fitur:
    - Authentication & Authorization (JWT)
    - Role-based Access Control (RBAC)
    - Achievement Management
    - Statistics & Reporting
    - User Management
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.prestasi.example.com/api/v1
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "student123"
          description: "Username atau email"
        password:
          type: string
          example: "password123"
          description: "Password user"
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: "login successful"
        data:
          type: object
          properties:
            token:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            user:
              $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          example: "student123"
        email:
          type: string
          example: "student@example.com"
        full_name:
          type: string
          example: "John Doe"
        role:
          type: string
          example: "student"
        permissions:
          type: array
          items:
            type: string
          example: ["achievements:create", "achievements:read"]

    # Achievement Models
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "invalid request"
        message:
          type: string
          example: "Detailed error message"
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Login dengan username/email dan password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh JWT token
      description: Refresh expired JWT token dengan refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  example: "refresh_token_here"
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "token refreshed successfully"
                  data:
                    type: object
                    properties:
                      token:
                        type: string
        '400':
          description: Invalid request
        '401':
          description: Invalid refresh token
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Logout user dan invalidate token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "logout successful"
        '401':
          description: Unauthorized

  # Profile
  /auth/profile:
    get:
      tags:
        - Authentication
      summary: Get user profile
      description: Mendapatkan profil user yang sedang login
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                  username:
                    type: string
                  email:
                    type: string

  # Users (Admin)
...
  /users/{id}:
    get:
      tags:
        - Admin
      summary: Get user by ID
      description: Admin mendapatkan detail user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
        '404':
          description: User not found

    put:
      tags:
        - Admin
      summary: Update user
      description: Admin mengupdate data user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User updated successfully

    delete:
      tags:
        - Admin
      summary: Delete user
      description: Admin menghapus user (soft delete)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted successfully
  /users/{id}/role:
    put:
      tags:
        - Admin
      summary: Update user role
      description: Admin mengubah role user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role_id
              properties:
                role_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Role updated successfully

  # Achievements
...